`Enumerator#map` with no block should return `Array[untyped]` Enumerator
According to the old RBS definition, the following code is considered to
return `Array[Integer]`:

```
[1,2,3].map.with_index {|x, _i| x.to_s }
```

... since `[1,2,3].map` returns `Enumerator[Integer, Array[Integer]]`
and `Enumerator#with_index` returns its second type variable.

However, it is obviously wrong since Ruby returns `Array[String]`.
It would be best if it can return `Array[String]`, but it seems hard to
represent it in the current RBS.

So, this changeset does next best; it makes `Enumerator#map` return
`Enumerator[Integer, Array[untyped]]`.  This is a bit unfortunate but
not wrong at least.