[docs] Add comments to sytnax doc
Dearest Reviewer,

I was shocked to see comments in rbs files only because it was not listed in the syntax file. I assumed that post line comments are not allowed. I am unsure if that is a correct statement.  I also assumed that heredocs/docstring/<<- and =begin/=end docs are not supported. 

Please let me know if I am incorrect in these assumptions and I will update the diff.

Thank you for work and time on this project,

Dictated but not reviewed,
Becker