Make Enumerable's predicate blocks return bool instead of untyped
For example, the return value of the block of Enumerable#all? is handled
as bool. In this case, it will bring no difference, but I think it is a
good habit to use a more explicit name instead of "untyped".

Note, if the definition of `bool` is changed according to #133, they
should be changed again.